fix: correct amount in float

+++ src/Pricing/Strategy/FlatRatePricingStrategy.php
@@ -13,7 +13,7 @@ final class FlatRatePricingStrategy implements PricingStrategyInterface
         return PricingModel::FlatRate === $pricingModel;
     }
 
-    public function calculate(Equipment $equipment, int $durationInDays): int
+    public function calculate(Equipment $equipment, int $durationInDays): float
     {
         $rates = $equipment->getPricingRates();
         $rate = $rates->first();
