+++ src/Pricing/Strategy/PricingStrategyInterface.php
@@ -9,5 +9,5 @@ interface PricingStrategyInterface
 {
     public function supports(PricingModel $pricingModel): bool;
 
-    public function calculate(Equipment $equipment, int $durationInDays): int;
+    public function calculate(Equipment $equipment, int $durationInDays): float;
 }
